Restructure core folder for better modularity and maintainability
- Reorganized files into structured folders within the core directory.
- Fixed imports across the project to align with the new structure.
- Included imports in `core/__init__.py` for better importability.
