+++ app/core/database/__init__.py
@@ -1,22 +1,8 @@
 from importlib import import_module
 from pathlib import Path
-from typing import Any
-
-import toml
 
 
-def get_version_from_pyproject() -> str:
-    """Retrieve the version value from the 'pyproject.toml' file."""
-    try:
-        data: dict[str, Any] = toml.load(Path("pyproject.toml"))
-        version = data["project"]["version"]
-        if not isinstance(version, str):
-            raise ValueError("Invalid version format in 'pyproject.toml'.")
-        return version
-    except (FileNotFoundError, KeyError, toml.TomlDecodeError) as err:
-        raise ValueError(f"Error reading version from 'pyproject.toml': {err}") from err
-
-
+# Automatically imports all model files to ensure Alembic can detect ORM models for migrations.
 def import_models_modules() -> None:
     """Imports all Python modules with 'model' in their filename."""
     # TODO: Improve the base path calculation. Using parent.parent.parent is not ideal.
