+++ app/core/gateway/handler.py
@@ -3,9 +3,9 @@ from fastapi.exceptions import RequestValidationError
 from starlette.exceptions import HTTPException
 from starlette.responses import Response
 
-from app.core.logger import main_logger
-from app.core.utils.logger_utils import prepare_log_message
-from app.core.utils.response_utils import render_error_response
+from app.core.config.logger import main_logger
+from app.core.gateway.error_response import create_error_response
+from app.core.gateway.error_response import prepare_log_message
 
 
 async def handle_internal_server_error(request: Request, exc: Exception) -> Response:
@@ -19,7 +19,7 @@ async def handle_internal_server_error(request: Request, exc: Exception) -> Resp
     else:
         error_context = f"An unexpected server error occurred. (Error type: {type(exc).__name__})"
 
-    return render_error_response(
+    return create_error_response(
         request,
         status_code=500,
         detail=f"Internal server error: {error_context}",
@@ -32,7 +32,7 @@ async def handle_validation_exceptions(request: Request, exc: Exception) -> Resp
         raise TypeError(f"Unexpected exception type: {type(exc)}. Expected RequestValidationError.") from exc
 
     main_logger.info(prepare_log_message("Client error - validation exception", exc, request))
-    return render_error_response(
+    return create_error_response(
         request,
         status_code=400,
         detail="Invalid input data. Please check your request.",
@@ -47,21 +47,21 @@ async def handle_http_exceptions(request: Request, exc: Exception) -> Response:
 
     if 400 <= exc.status_code < 500:  # Client error  # noqa: PLR2004
         main_logger.warning(prepare_log_message("Client error", exc, request))
-        return render_error_response(
+        return create_error_response(
             request,
             status_code=exc.status_code,
             detail=exc.detail or "A client error occurred.",
         )
     elif 500 <= exc.status_code < 600:  # Server error  # noqa: PLR2004
         main_logger.error(prepare_log_message("Server error", exc, request))
-        return render_error_response(
+        return create_error_response(
             request,
             status_code=exc.status_code,
             detail=exc.detail or "A server error occurred.",
         )
     else:  # Unhandled status codes
         main_logger.critical(prepare_log_message("Unhandled status code", exc, request))
-        return render_error_response(
+        return create_error_response(
             request,
             status_code=exc.status_code,
             detail=f"Unhandled status code: {exc.status_code}",
