+++ app/domain/user/services/auth/backend.py
@@ -4,9 +4,9 @@ from starlette.authentication import AuthCredentials
 from starlette.authentication import AuthenticationBackend
 from starlette.requests import HTTPConnection
 
-from app.core.config import app_config
-from app.core.database import get_session
-from app.core.utils.exceptions import AuthError
+from app.core.config.settings import app_config
+from app.core.database.session import get_session
+from app.domain.user.exception import AuthError
 from app.domain.user.model import User
 from app.domain.user.services.auth.models import AuthUser
 from app.domain.user.services.auth.security import pwd_context
