+++ app/routes/api/auth/user.py
@@ -5,7 +5,7 @@ from fastapi import Request
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.authentication import requires
 
-from app.core.database import get_session
+from app.core import get_session
 from app.domain.user.crud import UserRepository
 from app.domain.user.model import User
 from app.domain.user.schema import UserSchema
