+++ app/routes/api/blog/blog.py
@@ -5,7 +5,7 @@ from fastapi import Request
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.authentication import requires
 
-from app.core.database import get_session
+from app.core import get_session
 from app.domain.post.crud import PostRepository
 from app.domain.post.model import Post
 from app.domain.post.schema import PostSchema
